session_supervisor: carve routing.rs (pure move)
